pixel-art-maker/functions/requirements.txt
functions-framework
flask
Pillow